tmf: Make the basic query method multi-thread safe
Merge the functionality of queryState() into loadStateAtTime(). Now
loadStateAtTime() will return a List of ITmfStateInterval directly,
instead of keeping a single array internally and returning different
references of it.

If the caller wants to multi-thread its access to the returned list,
it's completely up to it now.

Also updated the Javadoc to reflect the new behavior.

Fixes bug #376426

Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>