lttng: Handle existing but invalid state index files
If a ".ht" file already exists we try to re-use it first. But if
this step fails, now we'll just scrap the existing file and build
a new one.

Other than the many possible error causes, this will be useful
for dealing with incompatible changes to the file format, for example.

Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>