tmf: Split StateHistorySystem into two interfaces
This will offer a query-only interface, which has no method
to write to or modify the state history in any way. This is
much safer, now that the views are starting to use it.

As part of the process, the history-less StateSystem is no
longer easily available. It was never much used in practice,
but if it's ever needed in the future it should be relatively
easy to bring it back.

Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>