tmf: Provide a special interval end time return value for views
The complete state model in TMF is based on the fact that states
end *before* the next state is started. This ensures we cannot
have 2 active states for a given attribute at a given timestamp.

However, this is not ideal for the viewers, because it leaves
gaps between intervals if we use those timestamps as-is. After
much head-banging, the simplest solution that was found was to
add a method in the ITmfStateInterval interface that viewers can
use to get their own version of the interval's end time.

Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>