tmf: Synchronize accesses to the history files
It seems the file channel positions are saved in each thread's
context on Linux, but not on Windows. This makes concurrent
queries to the state history very prone to breakage. This
small change makes sure a complete node is read before a
thread can be preempted.

In pratice the performance impact shouldn't be too big, since
both threads would be blocked by the same disk even if we
allowed concurrent access.

Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>