tmf: Use TMF signals to trigger the state history building
Instead of sending an event request on its own, the history
builder will now listen to the TMF signals to control the
sending and cancelling of the event requests.

This allows us to benefit from the coalescing, which makes
sure we don't read the trace more often than we need to.

Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>