lttng: Drop the current node arrays in the the event handler
Since we now use "common locations" shortcuts for commonly used
attributes in the tree, we don't really save much by caching
the quarks for all CPUs and all current running threads. This
was initally done as a performance optimization.

This commit drops the current*Nodes arrays that were populated
as new CPUs were seen while reading the trace. We will now get
the "current" pointers by doing standard state system queries.

If there is a performance impact, it's unnoticeable. I think it's
worth keeping this file as easy to understand as possible, since
it's often used as example for building an event handler.

Change-Id: Iedb3a13da837f3676c72f3c91edbd2e6960879a1
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>