tmf: Fold StateHistorySystem back into StateSystem
At first, a distinction was made to differentiate state systems
that would have a history backend, and those that would not.
In practice, history-less systems are never used, and wouldn't
even work since some recent refactoring.

If we ever need to implement a history-less statesystem, it would be
cleaner to simply use a dummy backend which discards all the
intervals, and returns null or similar on all past queries.

Change-Id: I5c04f9514236878f6f53190f862347a5b19308f4
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/7070