tmf: Dispose of the model in the statistics viewer
Fixes bug 388022

When a statistics view is disposed, it cleans the complete static model
that may still be in used by another statistics view. It is useless to
suppress the complete model in the view, since the viewers are the one
responsible to delete their own data.

Change-Id: I450b935cebf317c9afca24858aef41139bd055e3
Signed-off-by: Mathieu Denis <mathieu.denis@polymtl.ca>
Reviewed-on: https://git.eclipse.org/r/7407
Reviewed-by: Bernd Hufmann <bhufmann@gmail.com>
IP-Clean: Bernd Hufmann <bhufmann@gmail.com>
Tested-by: Bernd Hufmann <bhufmann@gmail.com>
Reviewed-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
IP-Clean: Alexandre Montplaisir <alexmonthy@voxpopuli.im>