ctf: Fix feature.xml to allow building with pde-build
When trying to build the CTF feature with pde-build, it currently
fails because of unsatisfied dependencies.

We need to:
- Remove the e.o.l.ctf.parser dep completely (the feature ships
  this plugin, it cannot depend on it before it's built!)
- Change the order of the plugins so .parser is built first, since
  the .core depends on it.

Side note, this .parser plugin is the only one (apart from the help
plugin) that does not get put into a jar when built. I really have
no idea why...

Change-Id: Id520ba5fe4b2aa064f0850748774e14a6b83e49b
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/7716
Tested-by: Hudson CI
Reviewed-by: Bernd Hufmann <bhufmann@gmail.com>
IP-Clean: Bernd Hufmann <bhufmann@gmail.com>
Tested-by: Bernd Hufmann <bhufmann@gmail.com>