tmf: Add an abstract class for the StateChangeInput
The state-change-input plugin needed by the state system is
currently offered through a IStateChangeInput interface.
However much of the code is common between implementations.
Furthermore, the LTTng kernel plugin nicely splits the work
in two threads, so it would be great to be able to easily
re-use that feature.

This introduces a AbstractStateChangeInput class which will
provide a good ground base for building a state input plugin.
It will also wrap the multi-thread part of the code, so other
inputs can benefit from it.

The interface remains available for more exotic use cases.

Change-Id: I8c9210af62b63b48ffcdc54b3da35df7754e1583
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/7798
Tested-by: Hudson CI