tmf: Add the Tmf- prefix to the state system interfaces
In preparation with them being exposed into ITmfTrace, the
interfaces to the state system should wear the proper Tmf-
prefix.

The read-only interface that most views and users will see has
been renamed to simply ITmfStateSystem. The builder interface
that is used by the trace for its construction is now called
ITmfStateSystemBuilder.

Change-Id: Id947bb4c933387f5346f78963247e7460e3982cc
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/7964