tmf: Statistics provider based on event requests
This patch adds a TmfEventsStatistics, which uses trace event
requests and counts the instances of each event types to
provide statistics information. This is useful for trace types
that cannot use the default implementation using the state
system (traces where events are not chronologically ordered,
for example).

The previous TmfStatistics has been renamed TmfStateStatistics,
to better differentiate between the two alternatives that are
now available.

To keep the request coalescing, the separate threading is now
done is a new ITmfStatistics.updateStats(). This allows the signal
handler code to send the queries, but return immediately and
spawn a thread that will wait for the requests to be done.

When the results are ready, they will be sent back to the view
using the new TmfStatsUpdatedSignal. This makes the handler for
the StateSystemBuildCompleted signal obsolete.

Change-Id: I5b6237aef7f5c7cb373c68e4624376146a6e5086
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/8376
Tested-by: Hudson CI
Reviewed-by: Francois Chouinard <fchouinard@gmail.com>
IP-Clean: Francois Chouinard <fchouinard@gmail.com>
Tested-by: Francois Chouinard <fchouinard@gmail.com>