tmf: Add dispose methods to the state and stats backends
When a trace/experiment is closed while a state history is
being built (either for storing kernel states or statistics),
the build should be cancelled as part of the trace's dispose().
This means the threads should be stopped cleanly, the files
should be closed, and any ongoing requests should be cancelled.

Change-Id: I0313ed37a927330a46b68024d1926922ddcaefb8
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/8491
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
IP-Clean: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Hudson CI