Fix clearing of histograms
When histograms are cleared, the full trace histogram does not clear its
range duration, and this allows it to still be drawn, however with the
obsolete duration the range box can be drawn at arbitrary location due
to overflow of x coordinates. The range duration and other
instance variables are now cleared.

When histogram no longer has data after it has been cleared, the
mouseHover will not set the Canvas tool tip text. It will still show the
tool tip text of the previous occurrence that mouse was hovered. The
tool tip text is now cleared.

Change-Id: I2bf26e5a6e75484496070c5f1842e71c0a1ebe7c
Reviewed-on: https://git.eclipse.org/r/8797
Tested-by: Hudson CI
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
IP-Clean: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>