Change LTTng help plugin generation
The generation of the LTTng help plugin used to be done from 
a wiki stored at Eclipse.org. This wiki matches the latest 
delivered source code. However, when developing new features 
or updating existing features it is required to document the 
new or updated feature. But this cannot be done on the wiki
web page. This causes that the writing of documentation 
is often postponed before the next official release.

This commit changes the way the help plugin is build. The source
wiki text file along with the relevant images as part of are 
stored as part of the git repository. The build-helper.xml was 
updated to build the help from the local wiki file. This allows 
the designer to update the documentation along with the feature
implementation. Note, that the local wiki can be edited with the 
MyLyn wiki text editor.

When a release to open source is done and these new or updated 
feature are part of it, then the open source wiki at Eclipse.org
needs to be aligned to the wiki which is stored in the git 
repository.

How to know if someone updated the open source wiki independent to
the one in the git repository? For that, the build-helper.xml file
also builds the help plugin files as it used to be done. After a 
Ant build with the build-helper.xml file the changes will show 
as changes in the git repository. Knowing what changed, it can be 
easily transfered to the local wiki stored in git.

Note that the legacy help files were removed from the repository.

Change-Id: Iba1b140a78a5e4aaa2b2a7ffbdb0cf852f9b6030
Reviewed-on: https://git.eclipse.org/r/8794
Tested-by: Hudson CI
Reviewed-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
IP-Clean: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Tested-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>