Fix histogram current event and time stamp format parsing
The parsing of date format is updated to use the time stamp format's
time zone.

DAY_OF_YEAR field is not copied from the reference time as setting it
also updates the MONTH and DATE, causing premature break out of the
field-copying loop.

HOUR_OF_DAY field is used to copy from the reference time instead of
the HOUR field, as the latter only uses a 12-hour clock.

The current event text control now uses the current event time value as
a reference time for parsing, instead of the trace start time. This
allows a time stamp format preference, which contains insufficient date
and time fields to fully represent the whole trace range, to be usable
in the text control. The non-visible date and time fields are assumed to
be set equal to those of the current event.

Change-Id: I25c67787c20aa26a0f267f8d460b441b74185122
Reviewed-on: https://git.eclipse.org/r/9051
Tested-by: Hudson CI
Reviewed-by: Bernd Hufmann <bhufmann@gmail.com>
IP-Clean: Bernd Hufmann <bhufmann@gmail.com>
Tested-by: Bernd Hufmann <bhufmann@gmail.com>