tmf: Make TmfEvent immutable
The immediate benefit of immutable events is that we can avoid
doing copies in many places in the framework (for indexing,
checkpoints, et al.). It also alleviates many potential
concurrency issues.

For specific sub-classes that need to have mutable (non-final)
fields, the recommended - hopefully transitional - approach is to
re-define the fields in question and override the getters to make
sure the extended class's versions are used.

This was done for CustomEvent in this patch.

Change-Id: I6a9ec635eaa68b03d1c5a6875f60ece7d47cb024
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/10123
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
IP-Clean: Patrick Tasse <patrick.tasse@gmail.com>