tmf: Add method to replace a statesystem's ongoing state
Replace the vestigial TransientState.changeOngoingStateInfo()
(which was never used so far) with a similar method that can
generate a current state from an array of intervals.

For now, this this potentially dangerous method is only exposed
to classes extending StateSystem.

This will be useful for the partial history, but also for anybody
wanting to do funky stuff with the state system.

Change-Id: Ieebe6dfceb57c1437d81fd9a97c6be10040afeb4
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/10760
Tested-by: Hudson CI
Reviewed-by: Bernd Hufmann <bhufmann@gmail.com>
IP-Clean: Bernd Hufmann <bhufmann@gmail.com>