lttng: Don't throw irrelevant exceptions in state system tests
Some tests expect exceptions, so these have to keep the "throws"
clause. For everything else, handle the excpetion in a try/catch
which will fail() the test if they happen.

This makes it simpler to extend the test cases, since the throws
may or may not be relevant in the subclass.

Change-Id: If71b351ba22bd88377120833eb83e9b78717d579
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/10795
Tested-by: Hudson CI
Reviewed-by: Bernd Hufmann <bhufmann@gmail.com>
IP-Clean: Bernd Hufmann <bhufmann@gmail.com>
Tested-by: Bernd Hufmann <bhufmann@gmail.com>