tmf: Add support for partial state system history
A partial history is like a shim between a trace's state system
and a "standard" history backend.

In a standard history, all state intervals are inserted into the
backend. With a partial history, only the intervals that cross
pre-determined checkpoints are inserted. To rebuild any state
between checkpoints, we load the previous checkpoint and re-read
the events in the trace between the checkpoint and the time of
the query.

Partial histories do not support single queries. While those
could be implemented by using a full-query, it seemed to cause
problems with range queries, where multiple state/event queries
happen in parallel (which is not so bad, range-query performance
would be absolutely disastrous!).

Change-Id: Iad709a00d630a40bf5a1c2e13d55d901f9a1eaa5
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/10990
Tested-by: Hudson CI
Reviewed-by: Bernd Hufmann <bhufmann@gmail.com>
IP-Clean: Bernd Hufmann <bhufmann@gmail.com>