tmf: Replace state value types with an enum
Originally, we specified the low-level byte to write to the
history tree file in ITmfStateValue. Now that we have multiple
possible state backends, it's much more convenient to use an
enum in the framework, and define the type<->byte conversion
inside the History Tree only.

Change-Id: I8bd694672256c15737e41813b3146b2f929d8c78
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/11065
Tested-by: Hudson CI
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
IP-Clean: Patrick Tasse <patrick.tasse@gmail.com>