tmf: Remove ITmfStateSystem#isLastAttribute()
This method was meant to be used as a kind of
"wasThisAttributeJustCreated()". But, if no new attrbiutes are
created, the last attribute would still return true, even if
it has been existing for a while!

So remove this potentially confusing method from the API
completely.

Change-Id: I8ac873ba18193e875806d4645e03dd228acd846c
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/12250
Tested-by: Hudson CI
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
IP-Clean: Patrick Tasse <patrick.tasse@gmail.com>