tmf: Use the signal's trace set in signal handlers
In relevant signal handlers (like traceOpened, traceSelected,...),
to verify if the trace is the target of the signal, use the new
TmfTraceManager.getTraceSet() method instead of relying on the
currently-active trace set.

In practice, the behavior should be the same, but it's a bit clearer
and cleaner this way.

Change-Id: Ied4405e5d3cfc051dcb43cc60837752c1f8889ea
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/12649
Tested-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
IP-Clean: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>