CTF : Handle traces with events without eventID
Bug 387039 : First, the eventID was set to null when the event was
created. We change it for -2 as default value when not set. Then, we
must change the way to look for packetIndexEntry to adapt it to the
particular case where there are many packetIndexEntry with the same
timestamp.

Change-Id: Ia3bf16cc1013e41c18b5305978d1b99c30624bdc
Signed-off-by: Jean-Christian Kouame
Reviewed-on: https://git.eclipse.org/r/12949
Tested-by: Hudson CI
IP-Clean: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Tested-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>