TMF: Add the concept of host id to a trace
This patch adds the concept of a host to a trace.  By default, the host is the
name of the trace itself (so 1 trace => 1 host).  For CTF traces, the clock
uuid is used to uniquely identify the host (boot session) the trace was taken
from.

Change-Id: Ie35725bc24be4ea5678c94d5ab5680e67892f4be
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/14052
Tested-by: Hudson CI
Reviewed-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
IP-Clean: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Tested-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>