lttng: Fix Resources view event list
With the fix to the event iterator (42acdc2), it is no longer required
to add 'null' events in the zoomed event list for all empty ranges to
hide underlying events in the full range event list. This simplifies the
building of event list in the Resources view.

If required, 'null' events are added at the start or end of the list to
properly identify the time range that the list represents.

Adds an interface to TimeEvent to split an event while keeping it's
class type and value.

Some trace specific fields are moved out of the generic TimeGraphEntry.

Change-Id: I6b3058227e64e456601b40fa2ea6c0afe3886f14
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/14326
Tested-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>