lttng: Adjust the file name of the RCP archives
Add the build date/time to the file name and remove the SNAPSHOT part
of the version number for (and only for) tracing.rcp.product. This will
be used as the versioning for builds made from the Git tree.

Also added the location of the archive files to the README.

Change-Id: I5a2e24fb1f82b0de05f36168bbbbeffec919e250
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/15109
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
IP-Clean: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>