tmf: Fix StringIndexOutOfBoundsException in time graph
The Utils.formatNs method is fixed to not throw an exception when the
parameter is Long.MIN_VALUE.

The TimeGraphViewer is modified to ensure the time range variables are
kept to 0 when there is no valid time range set (including when the
requested time range is invalid as in [+inf,-inf]). The range is
prevented from going out of bounds due to adjustment for the minimum
time interval.

The TimeGraphControl is modified to disable mouse time selection and
status line updates when the time range is 0.

Change-Id: I62dfe520913d160dac37526f1c02396c34f67830
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/15614
Tested-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
IP-Clean: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>