tmf: Rework test trace classes
Currently, test traces are accessed using array indices, which is simple
but scales very badly. It's also very confusing when writing a new trace
and wanting to decided which test trace to use.

Rework the test trace classes by replacing them with enum's that do the
same thing. Unfortunately, you cannot extend enums, so duplication of the
names is needed, but everything else is centralized in one place.

Change-Id: I6e598048531682428c1142c18d1760c21de5aeac
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/16143
Reviewed-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Tested-by: Hudson CI