ctf: Reformat and remove debugs in the CTF-Grammar
This is mostly a grammar reformatting.
My plan is to change some rules in the grammar.
I prefer to land the reformatting before playing with the grammar.

The 'debug_print' is useless in the grammar.
Antlr (and antlrworks) provides a really good debugger for debugging
this kind of grammar.

Change-Id: I6e590a168d2e5ff4502c36ea2fd8523b723d4399
Signed-off-by: Etienne Bergeron <etienne.bergeron@gmail.com>
Reviewed-on: https://git.eclipse.org/r/19483
Tested-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
IP-Clean: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>