tmf: Remove back-reference from HT-Node to HT-Tree
The HT-Node needs to know about the layout of the node (size, n-child)
but not about the tree in which it belongs to. Any back reference
is typically a bad design. It complicated unittests and maintenance.

  HT-Tree --->  HTNode--\
   \                     \
    \---------------------> HTConfig

This patch-set is one of multiples to refactor the HT-Tree API.
The API refactoring will lead us to a easier algorithmic refactoring.

Change-Id: Ifa11cec6a89b9b1c698e0a2b2b18ac5ec70a7db0
Signed-off-by: Etienne Bergeron <etienne.bergeron@gmail.com>
Reviewed-on: https://git.eclipse.org/r/19555
Reviewed-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
IP-Clean: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Tested-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>