ctf: modify data-structure in CTF-Trace holding streams.
CTF allows multiple events to have the same ID as long as
they are not part of the same stream.

Thus, a unique event is represented by <stream-id, event-id>
and not only by event-id.

In this case, the CTF-Trace contains multiple streams
but maps all event-id into the same map (without stream-id).

Back-references are always a bad design.
In a future patch-set, the plan is to remove
CTF-Trace from CTF-Stream.

Change-Id: If5fc50d8541b152938a96447087d88c78e35e959
Signed-off-by: Etienne Bergeron <etienne.bergeron@gmail.com>
Reviewed-on: https://git.eclipse.org/r/19619
Reviewed-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
IP-Clean: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Tested-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>