lttng: Handle threads in statedump events
In the statedump of a kernel thread (tid), we used to only check the
'ppid' field to get the paren thread. This works only for *processes*,
not for threads.

For threads we should use the 'pid' field to indicate which is the
parent TID.

Thanks to David Goulet for pointing this out!

Change-Id: I6b309d671ce43958ca9c0b43c5912877dca14e38
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/19728
Tested-by: Hudson CI
Reviewed-by: Florian Wininger <florian.wininger@polymtl.ca>