ctf: add support for growing streams
This patch adds the notion of the parser being in several states.
These states are detailed in the CTFResponse file.
The possible states are OK (we can keep reading),
WAIT (data will be available but is not yet),
FINISH (the trace has no more data),
ERROR (something happened that should not).

Note: this patch is non-trivial, even though it has been tested,
if there are regressions later, this patch is an interesting
candidate to check.

Change-Id: I2e6bd4bdc9e8e490ebb43910a749bcdca9754214
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/21057
Tested-by: Hudson CI
Reviewed-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
IP-Clean: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Tested-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>