tmf: Rework trace import wizard
- Use standard directory and file trees from the base import resource
wizard. This fixes the problem of UI freezes in the previous
implementation when setting the checked state of a large file system
tree. Instead of having on single tree for directories and files there
are 2 trees side-by-side, one for directories and one for the files of
the highlighted directory.

- Introduce TraceFileSystemElement class for the tree viewers
This class extends FileSystemElement that is used in the standard
tree viewers of base import wizard class and it supports lazy
loading of the children elements.

- Only import directories as traces if they are valid directory
traces. This avoids directories to be imported when they are selected
in the tree. Right now only CTF traces are considered for that. A
generic solution needs to implemented for that.

- Don't try to validate a trace when changing the selection state
and a trace type is selected. The validation of traces can be very
long if many files and directories are selected and we don't want
to do that for each selection change. The trace type validation
is now only done after pressing the finish button.

- Directories are validated first using trace types for directories
By doing that, valid directory traces are not considered for file
trace types

- Support for auto detection of trace types
When this option is selected the wizard tries to detect automatically
the trace type. For CTF traces (directories) this is working well.
For file traces additional code is needed to detect the valid trace
type.

- Possibly long validation and import operation can now be cancelled.

- New checkbox to import unrecognized traces or to omit them
During auto detection it might happen that a trace type is not found
for a given file. This new checkbox allows the use to import the
unrecognized traces or to omit them. This checkbox value is persisted
in the dialog settings and loaded when opening the wizard afterwards.

- Trace type selection combo box has now sorted entries

Change-Id: I4920ad1d1de40e7665f1394549d3ff945b91d870
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/21831
Reviewed-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
IP-Clean: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Tested-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Tested-by: Hudson CI