tmf: Add new interface to get pre-defined data types
This way, other trace classes may implement this interface and be
queried on the events it might contain. This will be useful for analyses
who can query traces to know whether or not they might contain the
events it needs to execute.

This can be used by many traces. The TmfEventTypeCollectionHelper
class should help extract useful information from the set of events.
At least until Java 8 comes out.

Change-Id: I1758c0d99b2b4d4b91f5b7fa3ee14b82c325dede
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/22937
Tested-by: Hudson CI