tmf: Fix exceptions in TmfTraceManagerTests
This Junit tests causes NullPointerExceptions and
IllegalStateExceptions in the TmfStatisticsModule and
TmfStateSystemAnalysisModule because the trace object is null. This
happens because the job that executes analysis was scheduled but
the acutal test case is finished. This patch fixes the time issue
in the test case and not the concurrency issue of these analysis 
modules.

Change-Id: Iead1d372e3596d0abb0e9857c56d0872228e4e22
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/22988
Tested-by: Hudson CI
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>
IP-Clean: Genevieve Bastien <gbastien+lttng@versatic.net>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
IP-Clean: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>