ctf/tmf: allow multiple traces to be open with name clashing events
This is a problem not yet encountered but that will be a showstopper.
If you have two traces with event definitions that have the same name
but different payloads the second one will override the first.

This means that if I have two traces with events that are named "bob"
and we have a state provider for each that looks up different fields
in "bob", one will no longer return the correct values and exceptions
will be thrown everywhere.

Change-Id: I05f9bc0fbbd67374fcf3680d75fcb769c4f32a3a
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/23285
Tested-by: Hudson CI
Reviewed-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>