tmf: Fix IndexOOB exception in HT node with concurrent read/write
Similar to bug #432191, but with the History Tree backend. It's possible
for an attribute to be created, sent to the backend, and written to a
node, all while a query is ongoing. (That's a race condition where you
race a Formula 1 against a turtle, but the turtle somehow arrives first).

Simply ignore those new attributes, since their value at time 't' would be
null anyway.

Change-Id: I072a08e84795a321c3b2bb393d9880ac4ba13a6a
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/24674
Tested-by: Hudson CI
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>