tmf: Fix deadlock in analysis module
Deadlock occurred when renaming a trace while its analyses are ongoing.

The refresh of supplementary folders is now done outside of the analysis
synchronization lock.

It is not necessary to trigger the refresh when cancel() is called,
since it will be triggered when the analysis execution is completed
(whether it was cancelled or not).

Change-Id: Id65fe9d57fadab1d17096d7e20bb15f43d30f1c7
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/25323
Tested-by: Hudson CI
Reviewed-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Tested-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>