tmf: Fix simple resource leak warnings
Many warnings left in ControlFlow and Resource Views, where they
get a reference to a (closeable) analysis module, and then the
compiler expects them to close it.

Since the view don't really use the module for anything other than
getting the state system, we could provide a utility class to get
the state system from an analysis module directly, without leaking
references. Shouldn't be hard to do, but outside of the scope of
this patch.

At some point we should revisit the ITmfTrace#getAnalysisModule(s)
methods. Maybe it is not such a good idea to expose closeable objects
publicly.

Change-Id: I0633d40260d0e2d37ad9da30fc6bb34d68cd5f38
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/26486
Tested-by: Hudson CI