tmf: Provide a static method to retrieve state systems
This new method in TmfStateSystemAnalysisModule will:

- Null check the analysis module, so callers don't all have to do it.
- Start the execution of the module, if required (calling schedule()
  after it has already been started is a no-op).
- Avoid exposing a ITmfAnalysisModule object, which is Closeable
  because it holds disposable resources.

This fixes a bunch of the new potential leaks warnings (actually, it
isolates them in one place). Eventually the analysis module API might
have to be reworked, to either never expose the analysis module objects
publicly, or to separate the "accessor" objects from the ones holding
the resources.

Change-Id: I693d233f02a8f53c49c5dc429a84c098bd35f004
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/26487
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Tested-by: Hudson CI