Downgrade to Tycho 0.20.0 to fix build
Tycho 0.21.0-SNAPSHOT seems to fragile right now for us
to use for the Luna release. We needed 0.21 to work around
a bug in IcedTea but we can work around this issue by setting
the SWT_GTK3=0 environment variable when executing the build
instead of having it in the pom.xml. This is less convenient
but more prudent in order to stabilize the build.

Change-Id: I33fc64552c73a0cd8f2eac12fa27afab6af92557
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/27213
Tested-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>