lttng: Correctly skip tests if test traces are missing
It used to be "unsafe" to do assumeTrue() in @BeforeClass methods,
but since Luna it seems to be the other way around: it skips
correctly when put in an @BeforeClass, but results in a failure
if it's within a @Before. Makes perfect sense!

Now add this to abstract and extending test classes, both with
different @BeforeClass, and you get a recipe for disaster.

Change-Id: I83670475241325f2e0277055fad83d8943927624
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/30298