pcap: use a larger trace for parser benchmarking
The previous trace contained only 640 events, so the parser finished parsing the whole file in a couple of microseconds. Thus, the variation of the results was a large percentage of the total CPU time, which is bad.

This new trace contains 100k events (and is totally anonymized), so the benchmark gives more precise results. This allows to really see the influence of future commits on the parsing speed.

Change-Id: I45900d5a25160d37ceb4dc6a06175d6a6b74f475
Signed-off-by: Vincent Perot <vincent.perot@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/30472
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>