tmf: lttngControl: show error output first when handling error
This make sure that event in machine interface mode the error are
positioned first and are visible. Mi generate a lot of data and can
saturate the information dialog.

This also cut the need to do after execution checkup and error processing into the
new MI service Controller. If in the future post processing is necessary
one could override the function and do the post-processing.

Change-Id: I96f4f1a97745ed6048a180bfde3681ebc1ac0351
Signed-off-by: Jonathan Rajotte Julien <jonathan.rajotte-julien@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/31395
Tested-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>