pcap: Convert filePath to java.nio.file.Path
The new Java 7 API for handling file paths is quite versatile,
and much more robust (cross-platform, etc.) than dealing with
simple Strings.

One note: I removed the check to verify that the files end in
".cap" or ".pcap". While it would still have been possible to
do with Path.toString(), on Unix typically the file contents
indicates the type, not the extension.

Change-Id: I327272dd99b298ccd61c1e53d4ded4ed79885d1c
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/31596
Tested-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-by: Vincent Perot <vincent.perot@ericsson.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>