lttng: Disable "Value of parameter is not used" warning
Recent versions of JDT in Eclipse 4.5 now include exception
parameters in the "value of parameter is not used" setting. This
throws upwards of 400 errors with our current settings.

Bug #439750 explains the problem. Bug #441933 tracks the addition
of a new UI preference to handle exception parameters separately
from method parameters.

For now, we can set the setting to Ignore to make Eclipse integration
builds useable. Once the two bugs above are fixed, we can revert
this commit.

Change-Id: I5db6322eb6d7ae3c2bb98ee983d0824e0e8fc9e2
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/33469
Reviewed-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Tested-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>