lttng: Do not use java.io.tmpdir property (/tmp) when possible
This is problematic when multiple instances of Eclipse are running
(builds).
Instead, we can create a "temporary" folder under the Eclipse workspace
so
that each Eclipse instances are independant. This also has the advantage
of
the directory being cleared every time a mvn clean is executed or when a
JUnit launch configuration is used inside Eclipse.

Change-Id: I3c7fb17d4b36e7ba5fe054ea9d81667c465e75b6
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/33728
Tested-by: Hudson CI