tmf: fix concurrency issue in analyses modules (bug 447434)
The access of member fTrace in class TmfAbstractAnalysesModule was
not thread-safe. The attribute was set to null when cancelling
the analyses (e.g. in the UI thread).

With this patch this attribute will be accessed (read and write)
using a synchronization lock and it will be checked for null
references before using it.

Additionally, this patch takes care of count down initialization
latches in case of cancellation.

Change-Id: Id632f128242ce85b04b7acdb79264ea466497947
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/35347
Reviewed-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>