ss: Also search in the ThreadedHistoryBackend's interval queue
If a view, like a live-enabled time graph view, makes a lot of queries
close to the "current end time" while the history is being built, it
is very possible to miss some intervals due to them being in the
ThreadedHistoryBackend's interval queue. We should also search in that
queue for possible intervals.

Since this case is still relatively rare, it's usually faster to query
the history tree first, then go look through the queue, rather than
checking in the queue first every time (which can only be explored
in O(n)).

Fixes bug #443127.

Change-Id: I82e389cf95ce04002f61629d5a0d317a35a931ff
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/35504
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Tested-by: Hudson CI
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>