tmf: Remove the precision concept from ITmfTimestamp
Timestamp precision is a problem in tracers, not in a viewer.
We already have the notion of scale which kind-of does the same
thing (if a timestamp is specified up to milliseconds, we can
consider it precise up to milliseconds).

Not only this removes big chunks of unused code, it should speed
up timestamp comparisons a little bit.

Change-Id: I55253721396cb05610503abd81861f522d3ee223
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/35426
Tested-by: Hudson CI
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>