tmf: Remove AutoCloseable from IAnalysisModule
While it was very useful for tests (every test manages its own module with
try-with-resources), in reality analysis modules are stored in the trace
but need to be exposed to external components. And the trace has no knowledge
of which types of modules it contains, so it cannot provide access methods. For
this reason it was more annoying than useful to have them AutoCloseable.

Change-Id: I62a1c588155c59a6a0047a52e2965e1f3acd5744
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/35425
Tested-by: Hudson CI