tmf: Use the DebugSuite test runner for some tests
We've had the DebugSuite for a while, which allows printing the name
of every test class being started. This is extremely useful for
debugging hanging tests when run with Maven.

We could now use this by default, so that the Hudson builds also print
that output. It will help pinpointing hanging tests there too.

The class is in tmf.core.tests, so for now only enable it for plugins
that depend on tmf.core.tests already. This is a good start, since so far
most hanging tests were in tmf.core.tests. We should move the suite to
the "common" package once we get around to creating that...

Change-Id: I775b8ed9b4b20962b0df9d1b4f84dd07556890d4
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/35498
Tested-by: Hudson CI
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>