tmf.ctf: Rename CtfIteratorManager classes
This patch simply renames existing constructs related to TMF-CTF iterators
to better represent what they do, to help future refactoring:

CtfIteratorManager -> CtfIteratorCEO .It is a "meta-manager" that manages
all the existing, now-called CtfIteratorManager's. Silly name, but don't
worry it's temporary ;)

CtfTraceManager -> CtfIteratorManager. This is the object that manages all
the iterators of one trace. It should eventually be moved to inside the trace.

Change-Id: Ie5483844856610907d3e332307eb0844209a93ef
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/35614
Tested-by: Hudson CI
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>