tmf: Remove source and reference from ITmfEvent
Those two notions were originally meant to have every trace type give
them a specific meaning, but to have some "common" concepts that
would be shown in the default event table.

Now that we have the possibility to aggregate columns in the event
table from multiple trace types, there is no need for those vague
concepts anymore. Every trace is free to implement and show any
specific item it wants. The base interface should be for the absolute
minimum that is required by the framework.

In fact, the notions of "source" and "reference" already meant different
things for different trace types (the Source of a CTF event was the source
CPU, while the source of a pcap event was the source address). So these
should *not* be aggregated in the same event columns since they do not
represent the same concepts.

Note: For the trace types that did make use of the source and reference
fields, this patch simply adds new fields to those traces, with the
names "source" and "reference" for now. But now that they will be decoupled
from the interface, we will be free to rename them accordingly.

Change-Id: I84f9729edee64233893779ca9c375a0a157bfc84
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/36131
Tested-by: Hudson CI
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>