lttng: Fix system calls for LTTng 2.6 kernel traces
The "syscall entry" and "syscall exit" events were renamed in LTTng 2.6.
This patch adds a new trace layout, which extends the base LTTng definitions,
but updates the system call event names.

"syscall exit" also had to be changed to a prefix, because LTTng 2.6 now lists
the syscall name in exits too. We don't make use of it though, so it shouldn't
affect other tracers who don't.

Also required updating the CFV/RV to only display the non-prefix part of the
syscall name, no matter what layout is used underneat.

Change-Id: I9061b0dff92daa11894d951c09ed3dffe80e4d1b
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/36199
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Hudson CI