lttng: Experimental support for perf-CTF traces
Implement initial support for perf kernel traces in CTF format.
They can re-use the event layout capabilities introduced previously,
so that we can add this support without implementing a completely new
trace type and accompanying analyses.

Also rename the "LTTng Kernel" trace type and analysis to "Linux Kernel",
as they are now tracer-agnostic.

The perf-in-CTF format is not set in stone yet, so the settings used
here will most probably have to be updated in the future.

Change-Id: I099f6e21caf40b9bfbe1a305c51947f1a264124b
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/36200
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Hudson CI