tmf: Execute TmfEventTable's request handler only on success
The handler for the TmfTimeSynchSignal in TmfEventsTable sends an event
request to "position" the experiment to the correct timestamp/event. This
request executes a seek() on the experiment itself, and is located in the
handleCompleted() callback.

It should be moved to the handleSuccess() callback instead: this
repositioning of the experiment should only be done when the request
finishes succesfully. It should not be run if the request was cancelled,
for example because the executor was flooded with requests and cancelled
some of them.

Partial fix for bug 451615.

Change-Id: I905bb5d19aa0fe40f9d63bfd2e4f60d90aba3050
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/36517
Tested-by: Hudson CI