rcp: Add a "deploy-rcp" profile to deploy nightly RCP builds
This can be added to the Hudson job to put nightly RCP builds
on download.eclipse.org.

It can be used locally (by overriding 'rcpDestination') but is not
of much use since it doesn't do much more than just copying files.
So I don't think it's worth documenting in the README.

Change-Id: Ic5db3f2fad8f9baec0e96b5d6c041ac09bd8ad9c
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/37309
Tested-by: Hudson CI
Reviewed-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Tested-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>