Re-enable the "unused method paramater" error
Revert of commit a6503a11.

Now that bug 439750 was fixed, and a new separate "unused exception parameter"
compiler option was added, we can re-enable the previously-used setting of
"unused method parameters".

This should not break existing 4.4 installations, since the old option didn't
check for exception paramaters at all. Only requirement would be for 4.5 users
to update to the latest integration build.

Only one such error slipped in while the option was disabled, which is honestly
not that bad!

Change-Id: I9d14daea3f004ed000ab9f9b72e46b5954579a8d
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/37892
Reviewed-by: Hudson CI
Reviewed-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Tested-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>