Support for upgrading Linux Tools features to Trace Compass
This patch modifies each feature's p2.inf files so that it can be upgraded from
the matching feature in Linux Tools.

The lines to add the Trace Compass repository are now only in the TMF feature
to avoid a lot of duplication and it is only necessary to put this in a "base"
feature that the user is likely to install. This is only a shorcut so that the
user doesn't have to add the update site manually.

Change-Id: I81a5e65c151995888df78f4b47282d6ad1cfcf06
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/39164
Reviewed-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>