Add SourceReferences to plugins
This allows the user to clone and checkout the source code of Trace Compass
without knowing where the repository is. See:
https://wiki.eclipse.org/PDE/UI/SourceReferences

One use case is to right-click on a plug-in under the Plug-in dependencies in
Package Explorer view and select Import from Repository. Note that for this
EGit needs to be installed.

What this patch does is to let Tycho generate the source references in the
MANIFEST.MF of the built jars. So to test this patch, one can build the update
site locally and inspect the manifests of the (non-source) jars.

Change-Id: Ie28c96a03df01fa9f15b13ccc6e0662f2cb3d0f4
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/39339
Reviewed-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>