rcp: Also deploy the RCP-specific update site
The RCP can use an update site separate from the main one for the plugin
version (that contains for instance the RCP product IU and features).
Populate this update site as part of the "deploy-rcp" profile.

Also removed the productDir property, this was used back when we tried
to define symlinks, which didn't really work out in the end.

Change-Id: I4a413164d4413a4767788a9efba938f98d6fdd77
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/40514
Reviewed-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Tested-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-by: Hudson CI