Add setup file for Oomph
This patch adds a setup file for Oomph. To test it:
1. Start the Oomph installer (https://wiki.eclipse.org/Eclipse_Oomph_Installer)
2. At the package selection, click the Advanced Mode button in the top-right
corner
3. Select Eclipse IDE for Eclipse Committers. Click next.
4. Select Eclipse.org
5. Click on the '+' button in the top-right corner, browse file system for
TraceCompass.setup.
6. Double-click on Trace Compass in the tree. Click next.
7. Enter installation folder name, workspace name. Click Next, Finish.

This should install and start a new Eclipse installation with the Trace Compass
code ready to be worked on. Once this patch is merged, the setup will be added
to Eclipse catalog so that the user doesn't have to browse for the file.

Change-Id: I3b52508531d604dc9b493b6b7d2386a617ba54c1
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/41099
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>