btf: Fix BTF event property source
The BTF UI plug-in is now started through the
org.eclipse.core.runtime.adapters extension, since no class would
trigger the plug-in start. This allows the BTF event adapter factory to
be registered. To force loading the plug-in, we let BtfEvent load the
adapter instead of simply getting it which in turn creates the factory
and loads the plug-in.

The BTF event property source is updated to display the Notes field into
its own property.

The BTF event reference member is renamed to target.

Change-Id: I37fe006cece41f9b470da714314cef932ab7e9a8
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/41302
Reviewed-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
(cherry picked from commit df854ddb27550ed4c040d5e4b2d978a117db24d9)
Reviewed-on: https://git.eclipse.org/r/41680