TMF: Have IAnalysisModule#setTrace return boolean instead of throw exception
In an eventual future where the appliesToTraceType will be deprecated, in favor
of using analysis requirements, the setTrace method returns a boolean to
express whether or not the trace was successfully set.

Consequence of this: the helpers can now explicitly return null if the trace
was not set correctly and the callers of the IAnalysisModuleHelper#newModule
must check nullness of the module.

Change-Id: Idb06b4a3b1741d345a3e3412feee84a98aa1fed0
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/41568
Reviewed-by: Hudson CI