tmf: Fix inconsistent tree item expansion in TimeGraphCombo
When refresh() is called, the tree viewer expanded all items but the
time graph viewer expanded to the auto-expand level. The tree viewer now
also expands to the auto-expand level.

When an entry filter is applied, the tree viewer expanded all items but
the time graph viewer expanded to the auto-expand level. The time graph
viewer now also expands all items.

Change-Id: Ied0d42d54a2a2b027e879db0d6f960854314eebe
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/42155
Reviewed-by: Hudson CI
Reviewed-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Tested-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>