ss: Move ownership of the SSID to the backend
Since the backend is the "deepest" element in the
state-system-stack, we normally have to create it first. And
to avoid back-references, other elements do not have access
to those above them.

The concept of State System ID however can be useful for any
component in the stack. For this reason, it's more convenient
to save it in the backend, where all components can access it.

Change-Id: I5dab7e358eccf7dee52ac7e5d2f8400e130d5161
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/42858
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-by: Hudson CI