tmf.ctf: Do not expose CtfIteratorManager publicly
The recent re-enabling of the API tools shows us that CtfTmfTrace
exposes the CtfIteratorManager type - which is internal - publicly.

We now get a hodgepodge of iterator methods in CtfTmfTrace, but
this is more or less unavoidable due to current constrains:
CTFTrace, CtfIterator and CtfIterator are internal, so cannot
be exposed publicly.

Change-Id: I2acf4727e78be238c149f2f93a8cf7bc19713f8e
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/43120
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>