Fix running tests with 4.4 target with 4.5 installation
The plugin org.eclipse.pde.junit.runtime bundled with Eclipse 4.5 is now
incompatible with Eclipse 4.4 because it requires org.eclipse.core.runtime
version 3.11.0 as the minimum which is only in 4.5. The first build that has
this problem is 4.5-I20150224-0800.
Trying to run tests would result in this error:

org.osgi.framework.BundleException: Could not resolve module: org.eclipse.pde.junit.runtime [217]
  Unresolved requirement: Require-Bundle: org.eclipse.core.runtime; bundle-version="[3.11.0,4.0.0)"

Also, we can point to the 4.4.2 update site which is faster than pointing
to the composite update site, especially when building with Tycho (dependency
resolution).

Change-Id: I910216656496cf961e580a69c78d285548d57452
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/43975
Reviewed-by: Hudson CI