ctf: remove variant tag definition from declaration
The tag definition of a variant was cached in the declaration.
This could create the possibility of a variant misreading the data.
The declaration is now more stateless and will not have these issues
pop up at a later date. Also this optimisation is no longer helpful
and consumes more memory.

Change-Id: I0136e936748ac5ea1e08d9b44caa0444fc6b6561
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/44184
Reviewed-by: Hudson CI
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>