Fix some instability in custom parser SWTBot tests
We need to wait until the XML files are written to disk after clicking on the
Finish button of the wizard. Otherwise, the content might be not be up to date
when we verify the content (assert) in the non-UI test thread.

From what I understand, SWTBotButton.click should be synchronous so all the
code in the UI thread should be done executing by the time it returns. But this
doesn't seem to be the case all the time. This will require more investigation
that might take quite some time so we can use this work around for now.

Change-Id: I9177b01f3ac425530aba454c26a97b5b7d31301d
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/44217
Reviewed-by: Hudson CI