control: Base code for profile dialog window
Modified ILttngControlService to add a loadSession command, also
implemented the methods in the appropriate children classes
(LTTngControlService and NullControlService).

Added ProfileDialog and ProfileHandler, as well as the appropriate
modifications to the plugin.xml to allow the execution of the handler as
well as the call to the Handler. As for now, pre-selected profiles must
be in
org.eclipse.tracecompass.internal.lttng2.control.ui/resources/folders.txt.
Appropriate .icons, strings and whatnot are externalized and at the
appropriate location > messages.properties.

Signed-off-by: Patrick-Jeffrey Pollo Guilbert <patrick.pollo.guilbert@gmail.com>
Signed-off-by: William Tri-Khiem Truong <tri-khiem-william.truong@polymtl.ca>
Signed-off-by: William Enright <william.enright@polymtl.ca>
Change-Id: Id312708b1907f3dd36fab65da7d96d5f917404a2
Reviewed-on: https://git.eclipse.org/r/45958
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/59118
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>