ctf: fix trace initialization bug
The trace when it is initialized will go to the second packet
since fPacketIndex is set to zero then incremented.

The test cases were updated because they were unfortunately made with
the assumption that the trace reader worked well without seeking to 0.

This bug does not affect trace compass, only third parties using the
CTF parser.

Change-Id: Ie38e2d53591fc0682b8aa339114cd3c3e197fc15
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/64589
Reviewed-by: Hudson CI