common: Introduce a generic BufferedBlockingQueue
The BufferedBlockingQueue (nicknamed "BBQ") is a wrapper around
a standard ArrayBlockingQueue, which accumulates elements in a
separate Collection before putting them in the actual Queue.
This allows a producing and a consuming threads to not block each
other as much when they are producing and consuming at the same
time.

An identified use case is the ThreadedHistoryTreeProvider, and
perhaps eventually the AbstractTmfStateProvider too.

Change-Id: I010547d7914c4c377bf1c85f4f830bff0aa8740f
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/46148
Reviewed-by: Hudson CI