common: Fix flaky BufferedBlockingQueue test
The test testing concurrent insertion/removals/iterations
incorrectly assumed that all elements send to the queue would
show up in the iterator.

This is a known limitation of BBQ's iterator(), and is documented
in its Javadoc. The elements present in the queue at the time the
iterator is created will be either seen by the iterator OR consumed
by the consumer.

We cannot assume anything with regards to the iterator's contents.
A separate (sequential) test for the iterator's contents will be
provied.

Change-Id: I3d281ca1ad4a81e69f02363b52942990424dd57c
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/47115
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>