common: Add external annotations for Map.get()/.remove()
Even if one declares a Map<Object, @NonNull Object>, Map.get()
can return a null to indicate "there is no value for this key".
Same applies to .remove() too.

Since the base methods are not annotated, we either have to
supply our own annotation, or make sure to always use @Nullable
keys. The former is less error-prone, so let's do that.

See http://stackoverflow.com/a/25371836/4227853 .

Change-Id: I3bb7395d7dca053d354f2f418d8529696edd0346
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/57249
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>