tmf: Fix IllegalStateException in HistogramDataModel
It should be allowed to have a last bucket set to 0 in the histogram
data model. This can happen if all trace events have the same timestamp.

The histogram time range is fixed to display even if the start and end
times are equal, as long as there is at least one event in the model.

The scaling is updated to make sure that these single-timestamp events
are spread over the whole width of the histogram. The scaled model will
have a bucket duration of 0 when it represents a single timestamp.

Change-Id: I5341aa6a158a3b4c1b3d4edee982ed67558e8a51
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/48224
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>