ctf: Remove the callsite support
The existing support for CTF callsites was based on an out-of-tree
branch of LTTng-UST. This branch was since then abandonned, and
callsites (now rather called "debug info") will be implemented
differently.

This makes CtfTmfEvent not extend ITmfSourceLookup, because for
all intents and purposes it should never have. Correct source
lookup will be added with the new support.

Change-Id: Ic8b9c133965a1230f7754bbdcb7bc933414f38b2
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/50310
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Hudson CI