tmf: Fix TmfEventField#equals()
Since the getter methods are not final, they can be extended
by sub-classes. This means the equals() and hashCode() method
should use these instead of the direct fields.

In the case of fFields however, since it is not exposed
directly, we can mark its separate getter methods final, and
use the class attribute in equals() safely.

We also need to handle the case where the Object value could be
implemented as a [] array of any type. This will require special
care, since standard Object.equals() will not check the contents.

Change-Id: I21f20ebee9a7c6485df9d16859f736b8ca961587
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/51213
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Hudson CI
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>