ctf: bug 471966: Fix file handle leak in CTF
If something goes wrong in the CTFStreamInputReader constructor or one
of its calls, the file handles in this class are not closed. This is
because the object is not fully constructed yet therefore the caller of
the constructor cannot close it.

Change-Id: Ie03164888711dd49d7b30ce5d774b4e4cdef9228
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/51426
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>
(cherry picked from commit 1ae2ec14818faa5df13d4d332919e5b07ef249d5)
Reviewed-on: https://git.eclipse.org/r/51506
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>