tmf: Fix flaky ImportAndReadSmokeTest test
In this test, it was possible for SWTBotUtils.openEditor to timeout
because it would not find the right Tree and it would wait for a
TreeItem forever. Instead, we can use the SWTBot from the view to narrow
down where the Tree is searched and it returns the right one. I saw this
test being flaky on Windows.

Change-Id: I44e98b6e88d795fd8b0d38b75ad0506e849af4b5
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/51535
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>