tmf: Fix FetchRemoteTracesTest on Mac
1) Opening the Preferences through the menu doesn't work...anymore?
I have gone back to Eclipse 4.4 and SWTBot 2.2.1 to see if it worked and it doesn't.
I suspect it might be a change in OSX but I don't have access to an older one.
I am also not 100% sure this ever worked but I thought it did.

2) We need to wait for the Preferences dialog to be active before looking
for the tree in it. That seems like a good fix for all platforms.

Change-Id: Ib9b3ed6edee75d368d3238451dc57a982f65ee0a
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/51861
Reviewed-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
(cherry picked from commit 78ef9da28ad51c95ea5a88f9ede978a6e2574e07)
Reviewed-on: https://git.eclipse.org/r/52815
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>