tmf: Add an abstract state system time graph view
This subclass of the abstract time graph view can be used when the time
events are built using a state system. The full states of the state
system are first queried chronologically for the whole time range using
a time resolution, and the list of full states is kept in memory and
reused by every time graph entry to build its time event list. This
avoids repeatedly loading the state system nodes from disk for each time
graph entry, which can be very slow if the full time range queries
require more nodes than is available in the state system cache.

Change-Id: I7dcac5d546a84462dba0e5a0cf320ba18d3617f5
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/52118
Reviewed-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>