ctf: Fix closed iterators still reattributed in CtfIteratorManager
CtfIterator extends CTFTraceReader which implements AutoCloseable,
making it closing its streams when used in a try-block.
CtfIteratorManager reattributes randomly the iterators when its
ArrayList contains the defined maximum number of elements. It thus
sometimes reattributes a CtfIterator with closed streams that
will never be reopened. This patch allows to verify if the
reattributed iterator has stream input readers before
reattributing it, and, if it's not the case, replace it with
a fresh new CtfIterator and report the error.

Change-Id: I67ff06280381529a01bac8dc0a596ca3e876754d
Signed-off-by: Raphaël Beamonte <raphael.beamonte@gmail.com>
Reviewed-on: https://git.eclipse.org/r/53275
Reviewed-by: Hudson CI
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>