ctf: Lose reference to StreamInputReader in EventDefinition
StreamInputReader is AutoCloseable, it is better to not have a
reference to it in the event.

Change the constructor to require the CPU number instead of the
whole StreamInputReader, as it is the only thing that the
EventDefinition uses.

While we are changing most calls, we can merge both versions of
the constructor into one, since they only differ by one @Nullable
parameter.

Change-Id: Ibeae8e7ff7e4f78ab1f11a24d61c739706bfc881
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/53327
Reviewed-by: Hudson CI