ctf: Fix leaking context in CtfTmfTrace.initTrace()
The context used to read the first event of the trace is never disposed.
Its trace reader is auto-closed after registering the event types.

This leaves the context with its closed trace reader in the iterator
manager. If the iterator manager becomes full, the trace reader can be
reused for another context but this will fail since it's been closed.

Change-Id: Iab224db1349d2ded5cb0f2d69b7b418a53e5eca0
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/53332
Reviewed-by: Hudson CI
Reviewed-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Tested-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/53516
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>