swtbot: Fix CopyToClipboardTest
The table sometimes did not react fast enough to the selection to remove
the Copy to Clipboard item from its context menu. The condition is
changed to wait until the item is no longer found and won't fail even if
the item is still found for a short time.

Also temporarily fixed to handle the consequences of SWTBot Bug 474063.

Change-Id: I6db683abbcec3ec86d4b70c67b6abcf588cd5783
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/53426
Reviewed-by: Hudson CI
Reviewed-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Tested-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>