tmf: Bug 474710: Time graph combo expanded state is not synchronized
When a time graph combo is refreshed, newly discovered items are
expanded to the auto-expand level in the time graph viewer, and the tree
viewer item expanded states are synchronized with those of the time
graph viewer. The expanded state of existing items is no longer affected
by the call to refresh().

Change-Id: I709681ef488e7d22991d5b7384c667532afd3410
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/53558
Reviewed-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/54017