tmf: Make StandardImportGzipTraceTest more robust
In a build, I got the failure:
"IllegalArgumentException: The row number (2) is more than the
number of rows(1) in the table"
http://eclip.se/5i

I think it's a case of the assertion hapenning before the table
was completely filled. Adding a few waitUntil should fix it.

Change-Id: I89b15a29bd0ae9f64f213fa0ffa715566127650a
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/53915
Reviewed-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
(cherry picked from commit 5daa4649921c6f99da5df9b74861589613835bb7)
Reviewed-on: https://git.eclipse.org/r/54020