ctf: make StreamInputPacketIndex more thread safe
This patch makes append synchronized. The SIPI can only grow and elements,
once in the index, cannot be changed. Therefore synchronization on reading
does not seem essential. If the same packet is written twice, the second
time, it will fail gracefully. The packet will still be there. This
should help "CtfTmfContextTest.testTooManyTraces".

Change-Id: Ic4dfcdda71fc1ebb58039066dab836e67458506f
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/54344
Reviewed-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>