analysis.linux: add TASK_STATE_DEAD and javadoc to LinuxValues.java
TASK_STATE_DEAD is useful to check if a task is still runnable or not.
This will be useful to see if a task can run or not when it is scheduled
out.

I had to read the kernel code for a while to figure out what
some of the flags mean, this should help others save time.

Change-Id: If538ce6b928c40ccf985ea655c58ce2da268539d
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/54630
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-by: Hudson CI