ctf: Introduce IPacketReader
Introduce the IPacketReader api to allow better decoupling between the
streaminputreader and the packet reader. This allows for more packet
reader types in the future, such as an InMemoryPacketReader for streamed
in RAM traces.

The IPacketReader is currently implemented by 2 classes:
    * a NullPacketReader, if there is no data to read
    * a CTFPacketReader, the typical packet reader

The CTFPacketReader is quite different from the CTFStreamPacketReader. The
SPR holds on to a FileChannel and maps segments as it needs. This results in
a fast but complex sliding window like reader. On the other hand, the
CTFPacketReader will only read one packet of a stream, a new one needs to
be created for each packet. This makes it wrap a single BitBuffer. As
FileChannels are resources but Bit(Byte)Buffers are not, this means a
CTFPacketReader is NOT autoclosable.

It should be noted that ByteBuffers can hold onto a system file handle. They
are released on GC. Also, ByteBuffers as they have their handle, will not
require the fileChannel to be opened all the time, so resources can be
dynamically closed when running low on file handles.

This patch also fixes some bugs in the tests relating to uninitialized traces
and lost event locations that were hidden due to "event.equals" not working as
well as it should.

The reason this patch does two things at the same time is that the interface
required a major re-write of the packet reader. This resulted in several bugs
being highlighted and fixed. To fix them in a separate patch before would
be very time consuming, and fixing them after would mean deliberately writing
a buggy packet reader to fix it in a subsequent patch.

This change will allow the following changes down the road:
    * In memory ctf reading (network streams)
    * Encrypted ctf stream reading
    * Compressed ctf stream reading
    * Multi-threaded stream reading
    * A power set of the previous elements

Change-Id: Id45b69bb0dac7a1f74f50cd732cd92064595f737
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/53284
Reviewed-by: Hudson CI