tmf: Fix Negative seek offset in indexer
If a Btree is saved to disk without any checkpoints stored in it, the
root node is not serialized out and invalid data is read when
reopening the index.

This is because the dirty flag for a node is set to false initially.
This works correctly for all child nodes because they are always
modified at least once right after being allocated but this is not
true for the root node since it is always allocated regardless of
whether or not it will contain checkpoints (and other child nodes).
Setting the dirty flag to true initially fixes the issue because
this will ensure nodes are serialized-out at least once. Nodes that
are serialized-in are not dirty until modified again.

Bug: 475543
Change-Id: I62db56445951f40ce0f27af04dd6f0d927f0c2fc
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/54841
Reviewed-by: Hudson CI