rcp: Use rcp repository as source for creating the installations
With this option, the build implicitly verifies that it would also
be possible to install the product (and its features) from that
repository.

This means that it provides some basic sanity check for installing
Trace Compass features. At least, for the ones included in the RCP.

For example, it would have uncovered the missing segement plugin in
the tmf feature, fixed here: https://git.eclipse.org/r/#/c/54804/

See
https://eclipse.org/tycho/sitedocs/tycho-p2/tycho-p2-director-plugin/materialize-products-mojo.html#source

Change-Id: I54e8ab63f830cbdb712b6c5c88d2c344f38e4518
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/54918
Reviewed-by: Hudson CI
Reviewed-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Tested-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>