tmf: Fix wrong context menu in ColumnHeaderMenuTest
This worked before because SWTBot didn't send the appropriate
MenuDetect when right-clicking on a table so the TmfEventsTable
still had the last context menu (the header's). Now that it does
send MenuDetect, it right-clicks in the middle of the table and the
context menu is not the header's context menu. The context menu call
should be done on the header instead.

Change-Id: I205d3f1de12d0b8b04100d956f5c1557aa10f57d
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/54940
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-by: Hudson CI
(cherry picked from commit e33ad2c1a850b33299a2910933c24f4821685922)
Reviewed-on: https://git.eclipse.org/r/54945