ctf: Add error logging when a closed iterator is selected for reattribution
In replaceRandomElement, a check has been added to create a new
iterator when encountering a closed one. This situation should
never happen, and so means that something is broken somewhere
else, e.g. that an ITmfContext hasn't been disposed properly.
This patch adds an error message log when that case is met to
allow further investigation and correction of the origin of
the problem.

Change-Id: Ib35f9906045f21042fbc9fea216a0e5da81221d3
Signed-off-by: Raphaël Beamonte <raphael.beamonte@gmail.com>
Reviewed-on: https://git.eclipse.org/r/55242
Reviewed-by: Hudson CI
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>