tmf: Avoid zoom on selection range update in AbstractTimeGraphView
When the selection range is updated externally, the selection is set on
the time graph widget with the 'ensure visible' flag set to true. This
may or may not cause the window range to be updated.

The AbstractTimeGraphView unconditionally starts a zoom thread in case
that the window range has changed. It should start it only if the range
has changed.

However since the fix to Bug 476148, if the selection update causes the
window range to change, the AbstractTimeGraphView will always be
notified through the ITimeGraphRangeListener callback, and this will
start a zoom thread. So it is now unnecessary to worry about starting a
zoom thread on handling of the TmfSelectionRangeUpdatedSignal.

Additionally, in the handling of this signal, the call to method
synchingToTime() is changed to use the time graph's selection time, not
the one received in the signal, as the time graph may clamp the
selection time according to the current time bounds.

Change-Id: I9e5d1f69107f56d4940fb2a952b41cb150832635
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/55670
Reviewed-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>