tmf: Clamp single time selection to time graph bounds
When the selection is updated to a single time, it should be clamped to
the current bounds. This will make it consistent with the handling of a
selection range update.

Additionally, when the selection range is updated with possible
notification, the check to see if the selection range has changed is now
made by comparing the clamped values, not the received values. Although,
in practice, since this update is triggered internally, it should not be
received with out-of-bounds values.

Change-Id: Iee93f7ef401fa6efc087aa03687de51f7f994590
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/55671
Reviewed-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>