analysis: Correctly track the state of scheduled-out processes
On a sched_switch, the state of a the process going out should not
always be WAIT_BLOCKED (the yellow state in the CFV). If the prev_state
field indicates 0 (TASK_RUNNING), it means the process was scheduled
out due to preemption, not to being blocked on something. In those
cases, we should put the process into the WAIT_FOR_CPU state.

Bug: 477052
Change-Id: Ic181811c11029d6374a5e65149a103febd25b5ea
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/55770
Reviewed-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/56223
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>