analysis: improve graph statistics scalability
The TmfGraphStatistics was keeping all duration values. Individual
values are not required to display summary statistics. Instead, this
patch updates the sum per IGraphWorker. Also, we introduce the fTotal
field and remove the special placeholder String key in the map, allowing
to rectify the keys type from Object to IGraphWorker.

* update v2:
  * synchronize only on fWorkerStats
  * make only one map lookup to update worker sum

Change-Id: I202945bdad6a0cb759b2fc4cbb791b3afef7c3cc
Signed-off-by: Francis Giraldeau <francis.giraldeau@gmail.com>
Reviewed-on: https://git.eclipse.org/r/55780
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>