tmf: Fix TimeGraphCombo item height synchronization issues
- Remove the caching of item height data in the tree item widget that is
meant to prevent repeatedly setting the time graph item height.

This only saved a hash map lookup in the time graph, but required two
linear searches in the TreeItem widget keyed data. It also could prevent
the height alignment to be performed if the time graph changed its item
heights without the time graph combo knowing about it. For example, all
time graph item heights are reset on refresh.

- Set the input on the time graph before setting the global item height.

Getting the global item height involves refreshing the list of visible
expanded elements, which also synchronizes their expanded state. If the
time graph input is not set yet, this causes all tree items to be
unnecessary (but temporarily) collapsed. As as counterpart to changing
the order, setting the global item height on the time graph will update
the item height of every existing time graph item. This only happens
once (when the input is set).

Change-Id: Id3519cbf8f9ce283583ac75af9c4f3ec7174b24a
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/55786
Reviewed-by: Hudson CI
Reviewed-on: https://git.eclipse.org/r/56013
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>