tmf: Fix incorrect TimeGraphFilterDialog result when dialog shell closed
The result is set to null when the Cancel button is pressed. But when
the dialog shell is closed by pressing the [x] widget, the result
remains set to the value set the last time that the dialog was closed
with a button. The first time that the dialog is opened, the result is
initialized to an empty array, so closing the shell filters all items.

Change-Id: I82225608e54a233edc9fa7ee66dd826c55d50c52
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/55999
Reviewed-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>