tmf: Fix for visible middle scrollbar on Mac
On Windows we need to use a workaround to hide the vertical
scrollbar. By using that workaround on Mac, it actually makes the
vertical scrollbar visible all the time. Instead, we can use the
SWT.H_SCROLL | SWT.NO_SCROLL combination to only get the horizontal
scrollbar. This doesn't work on Windows but can still be used on Mac
and Linux (GTK).

Bug: 466870
Change-Id: If199708b41ed1b9467c6950d5fcf7ce7e461f1e1
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/50115
Reviewed-by: Hudson CI
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>