tmf: Support markers in time graph control
The time graph can be provided with a list of markers.

Each marker has a time, duration, and color. The color can have an alpha
value for transparency. A marker can be specific to a particular time
graph entry, otherwise it is drawn for the full height of the control. A
marker can have an optional text label. A marker can be drawn in the
foreground (above the entry's time events) or in the background (below
the entry's time events).

The drawing of markers can be turned on or off globally.

Change-Id: I15d19925ab5371f21fbaf81d51182d17868e92a6
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/56472
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>