tmf: Add the possibility to choose the root node for synchronization
The root node in the spanning tree used for synchronization was
always the first hostId in lexicographic order.
Now the user can choose the trace that will be used as reference
for the synchronization.

When the user wants to synchronize an experiment a window similar
as the one used to add a trace to an experiment is shown.

The first trace is selected by default.

Change-Id: Iaaeb4c7905dc6171d9f8a40c4dbf4e8ba65d5ebd
Signed-off-by: Cédric Biancheri <cdc.biancheri@gmail.com>
Reviewed-on: https://git.eclipse.org/r/56769
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>