tmf: Fix time-axis alignment for cases where the series is not sorted
The time-axis alignment code for XY chart viewers (SWTCharts) assumes
that the series in the chart will be sorted and that the last item in
the series will be the one with the highest value. Since SWTChart
allows specifying series values in no specific order, this should be
handled properly in the time-axis alignment code.

We do not have any views right now that shows this behavior but if
you want to see this bug, you can try this patch in combination with:
https://git.eclipse.org/r/#/c/57011/

Change-Id: I8a850f9a12f945fe5616b0b82176d98ea45c33ae
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/57552
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Hudson CI