control: Bug 480217: Update MI XSD file for LTTng 2.7 targets
With the new XSD it is possible to use the Control view with LTTng 2.7
targets. All existing features that are supported in the Control view
right now will work again.

A basic test case was added to verify the XSD validation a basic lttng
session created with the Control view for a 2.7 target.

Change-Id: I72e4e89b6d461f0bfa45a13fb8dbf26a6a162223
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/58514
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>