common: Add external annotations for Guava's Table#get
Just like Map.get(), Table.get() can return a null value to
indicate that the key is not present in the table, independently
of the declaration of the type V.

Guava Tables are now used in the graph/critical path classes, so
this is also required.

Slightly harder to add annotations to Guava classes atm due to
https://bugs.eclipse.org/bugs/show_bug.cgi?id=480462
but one can still write the .eea files manually.

Change-Id: Id5cc1bdea73bafe62af60df0fdf7f3e9c5f40796
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/58776
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>