lttng: Do not throw exception if callsite binary is missing
If a trace with debug information is opened on a different machine
than the one where it was taken, it's very possible for the referred
binary files to not be present. In such a case it should not throw
an exception, but simply return that no source callsite information
is available.

The (upcoming) "binary callsite" information should still be available
in such cases though.

Change-Id: I5c3c77d9086b1445b8a1b11e8bb312929321d7f3
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/58883
Reviewed-by: Hudson CI
Reviewed-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Tested-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>