analysis: Improve comments for IKernelAnalysis.java
Add javadoc to IKernelAnalysis. The tracepoints used by the kernel state
system now are described in a simpler language, as it is unfair to assume
that an eclipse developper working with trace compass is also a Linux
kernel developer.

This patch is part of an effort to make it easier to adapt new operating
system traces to trace compass. The javadoc is written based on the commit
messages for given patches and Linux Kernel Maling List discussions as well
as wikipedia, trace metadata and Linux Weekly News updates.

Change-Id: I33eb53d0b172c3b9d235eaa2e4681eaa410e4e76
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/58886
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-by: Hudson CI