Latency: introduce latency density view
This patch introduces a view to view segment densities. The
view is based on SWTChart's bar graph. Its x axis is NOT
time aligned as it is not on the main timeline. It takes
the liberty to display time in a more "human readable"
form showing a scale and up to 3 decimals max. (1.234 ms)

There is a second part to this view: a selection table.
This table shows the selected segments from a drag operation.

Follow up tasks for this can be to introduce other filters and
improve the tooltips.

An example implementation is provided for the SystemCall analysis.

Change-Id: I238a61dc8eeeefdca0da107424686a64fb63d8c3
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/57573
Reviewed-by: Hudson CI
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>