Augment annotation to Map.entrySet()
We already annotated Map.entrySet() to return a

  @NonNull Set<Map.Entry<K, V>>

But actually, the returned Set should not contain any null elements,
since then entry objects will always exist, even if the key and/or
value is null. So we could augment the annotation to:

  @NonNull Set<@NonNull Map.Entry<K, V>>

K and V in this case need to remain unannotated, their type should
match the Map's declaration.

Change-Id: Id436fd3cfc4f679198070e5efdbb9d2d3ce478ae
Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-on: https://git.eclipse.org/r/59422
Reviewed-by: Hudson CI
Reviewed-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Tested-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>