tmf: Support adjustment of time graph window range on range selection
When the selection range is modified and 'ensureVisible' is true, then
the window range will be centered on either the selection begin time (if
it has changed) otherwise on the selection end time, if that time is
outside of the current window.

This adjustment is now enabled in time graph views when the
TmfSelectionRangeUpdatedSignal is received.

Change-Id: I8b5ec885b00009be05222db02acd745218759131
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/59758
Reviewed-by: Hudson CI
Reviewed-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Tested-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/61482
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>