tmf: Add Lost Events marker event source and factory
A marker event source factory for Lost Events is enabled for all traces.

The marker event source can reuse a previously returned marker list if
the requested parameters are identical to the previous request.

The event types statistics state system is updated to add a new
attribute which is used to compute the time ranges that are covered by
at least one lost event time range.

Change-Id: I2479afc3f07deb363f5847f4816940010bae3e9e
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/59389
Reviewed-by: Hudson CI
Reviewed-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Tested-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>