tmf: Fix FontEventEditorTest on Mac
On Mac, the Preferences menu item is under the application name.
For some reason, we can't access the application menu so we use the
keyboard shortcut. This commits create a helper method in SWTBotUtils so
that this problem doesn't occur again when a new test wants to open the
preference window.

Change-Id: Ib4a76cd6d0f1b70965f83eaf4920aa55cf81835a
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/60857
Reviewed-by: Hudson CI