Exclude Sonar rules that are not working with jdt.annotation
Since we have moved to jdt.annotation 2.0, a few Sonar rules (FindBugs)
do not work properly and report false positives. This resulted in
around 460 false issues in Sonar. We can disable those rules for now
until the problem is resolved. A bug report was filled at the FindBugs
project:
http://sourceforge.net/p/findbugs/bugs/1425/

The exclusions are done in the pom.xml so that any Sonar server can
benefit from this fix.

Change-Id: I31db810bc8a38d96870d94d5192139e6d74527b6
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/61292
Reviewed-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>