tmf: Try to make the shell fully visible in SWTBot at start
In certain environments, the workbench shell can be positioned at random
locations. Depending on the size of the display, it is possible that the
shell is not fully visible. This patch adds logic to the start of SWTBot
tests so that if the top shell is not fully visible, it will try to move
it in within the display bounds. If the shell cannot fit the display, it
will be positioned at the top-left corner of the display.

To be able to apply this to all test, a new utility method was
introduced, SWTBotUtils.initialize. More common code that is applicable
to all SWTBot tests could be moved there in the future.

This fix is important to some tests that can depend on whether of not
some controls get painted. For example, a drag listener is only added on
a sash when it gets first painted. If it is never painted,
TmfAlignTimeAxisTest fails when trying to drag this sash to verify the
proper behavior of time alignment.

Change-Id: Ic67739e204fcb0a4b567914e8b6078c286d56bd3
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/61094
Reviewed-by: Hudson CI
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/62023