tmf: Bug 477508: Fix wrong event count and ranks in TmfXmlTraceStub
The stub implementation of getNext() was increasing the context rank
twice per event. This would cause a wrong event count in the trace.

The processing of the 'set_aspects' dummy event was also causing the
context rank to be incremented twice. The XML Schema Definition is
changed to extract 'set_aspects' into its own element. The trace stub
now first parses the trace with a custom XML trace definition that only
processes the single 'set_aspects' event, then reinitializes the trace
with a custom XML trace definition that processes 'event' events.

Change-Id: Id1cf2251e364f892ab37511f0a5d99f2bcc2b5c6
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/61205
Reviewed-by: Hudson CI