common: Add a checkNotNullContents() for arrays
For Collections we can easily go through a Stream and back
to call checkNotNullContents(Stream<T>), for arrays it's slightly
more complicated, so we might as well define a new method
specifically for arrays.

Also add some "tests" (not actual runnable tests, but just snippets
of code that should compile) to make sure the annotations are working
correctly.

Change-Id: I2d18ba912076ea98cfca3628e7c8d6077209f74b
Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>
Reviewed-on: https://git.eclipse.org/r/59427
Reviewed-by: Hudson CI
Reviewed-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Tested-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>