Remove all "AllTests" suites
tycho-surefire-plugin's default configuration is to look for @Tests
in classes that start or end with "Test", or end with "TestCase".

We can make use of this and avoid all the nested AllTests.java suites
found in every single test package in the tree. This also avoids
the need for most custom configuration in each test plugin's pom.xml.

To run a complete test plugin from within Eclipse, one can still
right-click on the plugin itself, and choose
Run As -> Junit Plug-in Test.
(Doing this before this patch would actually run all tests twice!)

Note: this breaks the suites in the "alltests" plugin, but those could
be brought back by using something like ClasspathSuite, if desired.

Change-Id: Ibf960df37c5d3c98dc0e0dd722d7c9fe5c0b9d4f
Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>
Reviewed-on: https://git.eclipse.org/r/59442
Reviewed-by: Hudson CI
Reviewed-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Tested-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>