tmf: Fix for time-alignment bug
SashForm.setWeights() fails and throws runtime exceptions if
the sum of the passed values is 0 or if one of the values is
negative.

In some cases that include switching to a not-yet-initialized
Resource View, this situation could happen. Simply ignore sash
resize for invalid arguments. No need to time-sync an uninitialized
view anyway.

Change-Id: I3b7724795529ee91ae903ba3551491b5e0a7a645
Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/59881
Reviewed-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>