control: partly revert commit 0e7ea8ac and use clone in handlers
The patch that will be reverted here used copy constructors instead
of clone. However, using copy constructors the class type gets lost
in certain cases and this will cause different errors to happen.

For example, add context on a UST channel will send the command with
-k instead of -u. This is because a instanceof check for
ChannelCommandParameter fails since the object to check is a
CommandParameter instance and not ChannelCommandParameter.

I'll add a subsequent patch with a unit test for the add context case.

Change-Id: I01177bc1fad7eb100b204997b4572bf14cb5be61
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/61491
Reviewed-by: Hudson CI
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>